Move extensions to own directory
